Move roleToPermissionMap to shared
