+++ packages/back-end/test/models/dataSourceModel.test.ts
@@ -1,4 +1,4 @@
-import { Permissions } from "shared/permissions";
+import { Permissions, roleToPermissionMap } from "shared/permissions";
 import {
   updateDataSource,
   validateExposureQueriesAndAddMissingIds,
@@ -10,8 +10,8 @@ import {
 } from "back-end/types/datasource";
 import { testQueryValidity } from "back-end/src/services/datasource";
 import { usingFileConfig } from "back-end/src/init/config";
-import { OrganizationInterface, ReqContext } from "back-end/types/organization";
-import { roleToPermissionMap } from "back-end/src/util/organization.util";
+import { OrganizationInterface } from "back-end/types/organization";
+import { ReqContext } from "back-end/types/request";
 
 jest.mock("back-end/src/services/datasource");
 jest.mock("back-end/src/init/config");
